Add support for Telegram Bot API version 6.1

+++ src/types/video.rs
@@ -37,5 +37,5 @@ pub struct Video {
     pub mime_type: Option<Mime>,
 
     /// File size in bytes.
-    pub file_size: Option<u32>,
+    pub file_size: Option<u64>,
 }
